+++ CHANGES.rst
@@ -6,6 +6,8 @@ CHANGES
 Unreleased
 ----------
 
+- Add admonition box for display on documentation of older releases
+
 
 2022/03/01 0.21.2
 -----------------